+++ app/services/web_client.rb
@@ -1,3 +1,4 @@
 module WebClient
   class NotFound < Exception; end
+  class ContactAlreadyAdded < Exception; end
 end