+++ app/services/web_client/action_handler.rb
@@ -6,17 +6,29 @@ class WebClient::ActionHandler
 
   def initialize(nkey)
     @nkey = nkey
-    @notifications = Notification.find_notifications(nkey)
+    @notifications = Notification.by_nkey(nkey)
   end
 
   def do(action)
-    notifications.each do |notification|
-      notification.status = action && action.to_s
-      Notification::Save.new(notification).do
+    if already_added? && action == :ignored
+      fail(WebClient::ContactAlreadyAdded)
+    else
+      update_status(action)
     end
   end
 
   private
+
+  def already_added?
+    @notifications.first.added?
+  end
+
+  def update_status(status)
+    notifications.each do |notification|
+      notification.status = status && status.to_s
+      Notification::Save.new(notification).do
+    end
+  end
 
   def nkey_should_be_correct
     errors.add(:nkey, 'nkey is incorrect') if notifications.empty?