+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - :hammer: Fixed `ZazoActivity` score criteria (calculate all friends, raised ratio)
 - :hammer: Fixed `Contact::GetSuggestions` service to return only last non-friend contacts
 - :hammer: Changed database type of `notifications.compiled_content` from `string` to `text`
+- :hammer: Handle *ignore after adding* contact use case for `WebClient`
 - :hammer: Fixed minor bugs
 - :hammer: Refactored code